refactor: split PTY module into masker/sync/session + bug fixes
Restructure veil-cli for maintainability:
- src/pty/{masker,sync,session}.rs — PTY masking/sync/session
- src/commands.rs — all cmd_* functions
- src/proxy.rs — HTTP/HTTPS proxy
- veilkey_cli.rs 737→195 lines

Bug fixes from code review:
- session.rs: load_config(None) → load_config(patterns_file)
- session.rs: static mut → AtomicI32 (UB removal)
- session.rs: UTF-8 byte slicing → ceil_char_boundary
- sync.rs: long poll timeout 10s → 45s
- api.rs: legacy fallback enrich_mask_map missing
- api.rs: encoded variant dedup + URL-safe b64
- commands.rs, detector.rs: preview slicing UTF-8 safe
- masker.rs: line-clear false positive fix
- detector.rs: find_iter+re-captures → captures_iter
- logger.rs: count() JSON parse → line counting
- Remove unused portable-pty dependency
- Apply cargo fmt formatting fixes
