+++ services/installer/src/platforms/veil_cli.rs
@@ -15,7 +15,10 @@ pub fn run(args: VeilCliArgs) -> Result<(), String> {
     println!("  URL: {}", args.url);
     println!();
 
-    check_cmd("cargo", "Install Rust: curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs | sh")?;
+    check_cmd(
+        "cargo",
+        "Install Rust: curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs | sh",
+    )?;
     println!("{}", "[1/1] Prerequisites OK".green());
 
     // Delegate to bash script
@@ -33,7 +36,12 @@ pub fn run(args: VeilCliArgs) -> Result<(), String> {
 }
 
 fn check_cmd(name: &str, help: &str) -> Result<(), String> {
-    if Command::new("which").arg(name).output().map(|o| o.status.success()).unwrap_or(false) {
+    if Command::new("which")
+        .arg(name)
+        .output()
+        .map(|o| o.status.success())
+        .unwrap_or(false)
+    {
         Ok(())
     } else {
         Err(format!("{} not found.\n  {}", name, help))
