+++ services/veil-cli/Cargo.toml
@@ -33,7 +33,6 @@ urlencoding = "2"
 ureq = { version = "2", features = ["json"] }
 rustls = { version = "0.23", features = ["ring"] }
 chrono = { version = "0.4", features = ["serde"] }
-portable-pty = "0.8"
 libc = "0.2"
 nix = { version = "0.29", features = ["term", "signal", "process"] }
 base64 = "0.22.1"
