+++ services/veil-cli/src/detector.rs
@@ -181,12 +181,10 @@ impl<'a> SecretDetector<'a> {
         let has_context = self.has_sensitive_context(line);
 
         for pat in &self.config.patterns {
-            for m in pat.regex.find_iter(line) {
-                let full_match = m.as_str().to_string();
+            for caps in pat.regex.captures_iter(line) {
+                let full_match = caps.get(0).unwrap().as_str().to_string();
                 let value = if pat.group > 0 {
-                    pat.regex
-                        .captures(m.as_str())
-                        .and_then(|c| c.get(pat.group))
+                    caps.get(pat.group)
                         .map(|g| g.as_str().to_string())
                         .unwrap_or_else(|| full_match.clone())
                 } else {
@@ -255,9 +253,9 @@ impl<'a> SecretDetector<'a> {
                     .then(b.full_match.len().cmp(&a.full_match.len()))
             });
 
-            let mut replaced: HashMap<String, bool> = HashMap::new();
+            let mut replaced: std::collections::HashSet<String> = std::collections::HashSet::new();
             for det in detections {
-                if replaced.contains_key(&det.value) {
+                if replaced.contains(&det.value) {
                     continue;
                 }
                 if let Some(vk) = self.issue_veilkey(&det.value) {
@@ -267,10 +265,16 @@ impl<'a> SecretDetector<'a> {
                     } else {
                         line = line.replacen(&det.value, &vk, 1);
                     }
-                    replaced.insert(det.value.clone(), true);
+                    replaced.insert(det.value.clone());
                     self.stats.detections += 1;
-                    let preview = if det.value.len() > PREVIEW_LEN {
-                        format!("{}***", &det.value[..PREVIEW_LEN])
+                    let preview = if det.value.chars().count() > PREVIEW_LEN {
+                        let end: usize = det
+                            .value
+                            .char_indices()
+                            .nth(PREVIEW_LEN)
+                            .map(|(i, _)| i)
+                            .unwrap_or(det.value.len());
+                        format!("{}***", &det.value[..end])
                     } else {
                         "***".to_string()
                     };
@@ -290,8 +294,13 @@ impl<'a> SecretDetector<'a> {
                 if line.contains(&value) {
                     line = line.replace(&value, &vk);
                     self.stats.detections += 1;
-                    let preview = if value.len() > PREVIEW_LEN {
-                        format!("{}***", &value[..PREVIEW_LEN])
+                    let preview = if value.chars().count() > PREVIEW_LEN {
+                        let end: usize = value
+                            .char_indices()
+                            .nth(PREVIEW_LEN)
+                            .map(|(i, _)| i)
+                            .unwrap_or(value.len());
+                        format!("{}***", &value[..end])
                     } else {
                         "***".to_string()
                     };
